Rename file with constants
