+++ src/reducers/index.js
@@ -8,7 +8,7 @@ import {
   SET_CORNER_DOWN,
   SET_DIVIDER_DOWN,
   SET_STATE
-} from '../constants/BlenderLayoutConstants';
+} from '../constants';
 
 import {
   split,
